dashboard/requirements.txt
streamlit
pandas
kafka-python-ng